% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlssReport.R
\name{extract_terms}
\alias{extract_terms}
\title{Aux extract function}
\usage{
extract_terms(object)
}
\arguments{
\item{object}{An object of class \code{"gamlss"}.}
}
\description{
aux function used by gamlssReport; not intendent to be used by the user.
}
\examples{
library(gamlss)

data(aids)
aids1<-gamlss(y~pb(x,df=4)+qrt,data=aids,family=PO)
extract_terms(aids1)
}
\seealso{
\code{\link{gamlssReport}}, \code{\link{print.gamlssReport}}
}
\author{
Rok Blagus, \email{rok.blagus@mf.uni-lj.si}
}
