% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlssReport.R
\name{gamlssReport}
\alias{gamlssReport}
\title{Extract information from the GAMLSS fitted objects}
\usage{
gamlssReport(object)
}
\arguments{
\item{object}{an object of class \code{"gamlss"}.
Note: only one smooth term per parameter is assumed (this smooth term had to be obtained using pb()).}
}
\description{
The function extracts the paramaters of the fitted GAMLSS object.
}
\examples{
library(gamlss)
library(splines)

data(aids)
aids1<-gamlss(y~pb(x,df=4)+qrt,data=aids,family=PO)
m.aids1<-gamlssReport(aids1)
m.aids1

data(abdom)
mod<-gamlss(y~pb(x),sigma.fo=~pb(x),family=BCT, data=abdom, method=mixed(1,20))
m.mod<-gamlssReport(mod)
m.mod
}
\seealso{
\code{\link{plot.gamlssReport}}, \code{\link{print.gamlssReport}}, \code{\link{predict.gamlssReport}}, \code{\link{centile.gamlssReport}}, \code{\link{score.gamlssReport}}
}
\author{
Rok Blagus, \email{rok.blagus@mf.uni-lj.si}
}
