% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlssReport.R
\name{make_inverse}
\alias{make_inverse}
\title{Inverse of the link function}
\usage{
make_inverse(f, x)
}
\arguments{
\item{f}{the name of the link function (a character).}

\item{x}{the values at which to evaluate the inverse of the link function.}
}
\description{
Inverse of the link function
}
\examples{
make_inverse("log",1)
}
\seealso{
\code{\link{predict.gamlssReport}}
}
\author{
Rok Blagus, \email{rok.blagus@mf.uni-lj.si}
}
