% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlssReport.R
\name{print.gamlssReport}
\alias{print.gamlssReport}
\title{Print function}
\usage{
\method{print}{gamlssReport}(x, digits_range = Inf, digits_coefs = Inf, digits_knots = Inf)
}
\arguments{
\item{x}{an object of class \code{"gamlssReport"} obtained as a call to \code{\link{gamlssReport}}.}

\item{digits_range}{a numeric value determining how many digits to display for range of x; defaults to \code{Inf} (no rounding).}

\item{digits_coefs}{a numeric value determining how many digits to display for coefficients (fixed and penalized); defaults to \code{Inf} (no rounding).}

\item{digits_knots}{a numeric value determining how many digits to display for the knots of the B-spline basis; defaults to \code{Inf} (no rounding).}
}
\description{
Prints the extracted information that can be used in the publications
}
\examples{
library(gamlss)


data(aids)
aids1<-gamlss(y~pb(x,df=4)+qrt,data=aids,family=PO)
m.aids1<-gamlssReport(aids1)
print.gamlssReport(m.aids1)

data(abdom)
mod<-gamlss(y~pb(x),sigma.fo=~pb(x),family=BCT, data=abdom, method=mixed(1,20))
m.mod<-gamlssReport(mod)
print.gamlssReport(m.mod)
}
\seealso{
\code{\link{gamlssReport}}, \code{\link{plot.gamlssReport}}
}
\author{
Rok Blagus, \email{rok.blagus@mf.uni-lj.si}
}
