% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlssReport.R
\name{score.gamlssReport}
\alias{score.gamlssReport}
\title{Calculate the score for given centiles}
\usage{
score.gamlssReport(object, centile, newdata, gamlss.prediction = FALSE)
}
\arguments{
\item{object}{the object of class \code{"gamlssReport"} generated by \code{\link{gamlssReport}}.}

\item{centile}{the centile for which to calculate the score.}

\item{newdata}{a dataframe containing the Xs for which to evaluate the centile.}

\item{gamlss.prediction}{logical, if \code{TRUE} it uses the spline interpolation approach implemented in \code{predict.gamlss}, otherwise it uses the interpolation of the B-spline basis and the estimated penalized coefficients. Defaults to \code{FALSE}. Note: if set to \code{TRUE}, the \code{\link{gamlssReport}} had to be used with \code{extract.smooth=TRUE}, oterwise the error is returned.}
}
\value{
a vector containing the scores
}
\description{
Using the object generated by \code{\link{gamlssReport}}, the function calculates the score for a given centile and x
}
\details{
\code{centile} can be a vector, in this case \code{newdata} need to be a dataframe with single row, or dataframe with the same no of rows as \code{centile}; if \code{centile} is a scalar \code{newdata} can have as many rows as desired; see examples.
}
\examples{
library(gamlss)


data(aids)
aids1<-gamlss(y~pb(x,df=4)+qrt,data=aids,family=PO)
m.aids1<-gamlssReport(aids1)
score.gamlssReport(m.aids1,centile=c(0.1,0.5,0.9),newdata=data.frame(x=4,qrt=4))
score.gamlssReport(m.aids1,centile=0.1,newdata=data.frame(x=c(4,5),qrt= c(4,2)))
score.gamlssReport(m.aids1,centile=c(0.1,0.9),newdata=data.frame(x=c(4,5),qrt= c("reference",2) ))
}
\seealso{
\code{\link{gamlssReport}}, \code{\link{predict.gamlssReport}}
}
\author{
Rok Blagus, \email{rok.blagus@mf.uni-lj.si}
}
