% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlssReport.R
\name{ShinyApp.gamlssReport}
\alias{ShinyApp.gamlssReport}
\title{Generate Shiny app}
\usage{
ShinyApp.gamlssReport(
  obj,
  xname,
  range.x,
  plotArgs = "seq.length=1e3, xlab=xname, ylab='y'",
  dir = paste0(getwd(), "/ShinyApp"),
  run = TRUE
)
}
\arguments{
\item{obj}{An object of class \code{"gamlssReport"} representing the GAMLSS model summarized by the function \code{\link{gamlssReport}}.}

\item{xname}{name (a character of length one) of the x variable used on the x-axis of the plot.
Note, the name must exactly match the name of the variable that was used when fitting GAMLSS.}

\item{range.x}{a numeric vector of length 2 specifying the range of x for which to show the centiles.}

\item{plotArgs}{verbatim text (a character of length one) added after the first 3 parameters (i.e. after \code{obj,xname,range.x}) of \code{\link{plot.gamlssReport}}. Defaults to \code{xname=x, range.x=range(x), seq.length=1e3, xlab=xname, ylab='y'}.
Note: as least \code{"xname"} and \code{"range.x"} should be provided as they have no default.}

\item{dir}{(a character of length one) A directory where the function will save \code{app.R} and \code{gamlsReportObj.RDS} (a RDS file of an \code{obj}). Defaults to \code{paste0(getwd(),"/ShinyApp")}, i.e. ShinyApp folder in working directory.}

\item{run}{logical Run app after it's generation? Defaults to \code{TRUE}.}
}
\description{
Function \code{ShinyApp.gamlssReport} generates Shiny app, which plots centile curves for the object created by \code{\link{gamlssReport}}.
}
\details{
Function generates a new directory in which it saves two files: \code{"app.R"} and \code{"gamlsReportObj.RDS"}. When app is run it reads RDS file into gamlssReport object (copy of \code{obj} argument) and plots centile curves by using \code{\link{plot.gamlssReport}}. If user enters (x, y) or (x, centile) value pair, a dot (x,y) is added to the plot and a short comment is shown below the plot.
}
\examples{
library(gamlss)
data(abdom)
mod<-gamlss(y~pb(x),sigma.fo=~pb(x),family=BCT, data=abdom, method=mixed(1,20))
m.mod<-gamlssReport(mod)
## Not run:
#ShinyApp.gamlssReport(obj=m.mod, xname="x", range.x=c(10,45), dir="/TestDir",
#  plotArgs="seq.length=1e3, xlab='gestational age', ylab='abdominal circumference'")
## End(Not run)
}
\seealso{
\code{\link{gamlssReport}}, \code{\link{plot.gamlssReport}}
}
\author{
Bojan Leskosek, \email{bojan.leskosek@fsp.uni-lj.si}
}
