% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlssReport.R
\name{gamlssReport}
\alias{gamlssReport}
\title{Extract information from the GAMLSS fitted objects}
\usage{
gamlssReport(object, extract.smooth = FALSE)
}
\arguments{
\item{object}{an object of class \code{"gamlss"}.}

\item{extract.smooth}{logical indicating if the prediction for the smooth terms should be extracted (see details). Defaults to \code{FALSE}.
Note: multiple smooth terms per parameter are allowed (these smooth terms had to be obtained using \code{pb()}).}
}
\value{
An object of class \code{"gamlssReport"} for which \code{print}, \code{plot}, and \code{predict} functions are available.
A list with components
\itemize{
\item{\code{family}} {the distribution family used when fitting GAMLSS}
\item{\code{params}} {the names of the fitted parameters, i.e. mu, sigma, nu, tau}
\item{\code{link}} {the link function for the model used for each parameter (a named list, with names corresponding to the parameters)}
\item{\code{coef.beta}} {the linear coefficients for each parameter (a named list, with names corresponding to the parameters)}
\item{\code{coef.spline}} {the penalized coefficients for the P-spline for each parameter where pb was used to model non-linear association (a named list, with names corresponding to the parameters).}
\item{\code{knots.spline}} {the knots used to form the B-spline basis for the P-spline (a named list, with names corresponding to the parameters)}
\item{\code{range.x}} {the range of variable that was used for P-spline (a named list, with names corresponding to the parameters; \code{NULL} for parameters where \code{pb()} was not used)}
\item{\code{degree.spline}} {the degree of the spline (a named list, with names corresponding to the parameters; \code{NULL} for parameters where \code{pb()} was not used)}
\item{\code{terms}} {details about the fixed effects and nonlinear term}
\itemize{
\item{\code{fixformula}} {formula for the linear effects (a named list, with names corresponding to the parameters)}
\item{\code{splinevar}} {names of the variables that were used for P-splines (a named list, with names corresponding to the parameters; \code{NULL} for parameters where \code{pb()} was not used)}
}
\item{\code{term.smooth} if \code{extract.smooth=TRUE} the fitted smooth terms for the observed datapoints}
}
}
\description{
The function extracts the paramaters of the fitted GAMLSS object.
}
\details{
In the presence of additive terms there are two approaches for making interpolation (for the observed data points the two approaches return the same result): 1) interpolating the values of the B-spline basis for the new datapoints and using the estimated penalized coefficients to obtain predictions and 2) spline interpolation is performed using the estimated smooth terms for the observed datapoints. The latter approach (implemented in \code{\link{predict.gamlss}} using natural splines) requires that the fitted smooth terms and original datapoints for the variable(s) used to form the spline are saved. Setting \code{extract.smooth=TRUE} will save these and enable to exactly replicate the results of \code{\link{predict.gamlss}} when using \code{\link{predict.gamlssReport}}. Note that this will enable sharing the potentially sensible information!
}
\examples{
library(gamlss)


data(aids)
aids1<-gamlss(y~pb(x,df=4)+qrt,data=aids,family=PO)
m.aids1<-gamlssReport(aids1)
m.aids1

data(abdom)
mod<-gamlss(y~pb(x),sigma.fo=~pb(x),family=BCT, data=abdom, method=mixed(1,20))
m.mod<-gamlssReport(mod)
m.mod
}
\seealso{
\code{\link{plot.gamlssReport}}, \code{\link{print.gamlssReport}}, \code{\link{predict.gamlssReport}}, \code{\link{centile.gamlssReport}}, \code{\link{score.gamlssReport}}
}
\author{
Rok Blagus, \email{rok.blagus@mf.uni-lj.si}
}
