% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlssReport.R
\name{make_spline}
\alias{make_spline}
\title{Aux function to recreate the B-spline basis}
\usage{
make_spline(newx, knots, degree, rng)
}
\arguments{
\item{newx}{vector containing values of x for which the B-spline bases are to be calculated.}

\item{knots}{the vector of knots; currently only the length(knots) matters (ie we assumed that knots in call to GAMLSS were uniformly set)}

\item{degree}{the degree of the spline used in pb().}

\item{rng}{vector with 2 elements, min(x) and max(x)}
}
\description{
aux function used by predict.gamlssReport; not intendent to be used by the user.
}
\examples{
library(splines)

make_spline(10,seq(from=10,to=20,by=1),3,c(9,21))
}
\seealso{
\code{\link{predict.gamlssReport}}
}
\author{
Rok Blagus, \email{rok.blagus@mf.uni-lj.si}
}
