% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gamlssReport.R
\name{predict.gamlssReport}
\alias{predict.gamlssReport}
\title{Predict function}
\usage{
\method{predict}{gamlssReport}(object, newdata, ...)
}
\arguments{
\item{object}{the object of class \code{"gamlssReport"} generated by \code{\link{gamlssReport}}.}

\item{newdata}{A dataframe containing all the variables (the names must match comepletely) needed to make predictions; see details for factors.}

\item{...}{for extra arguments.}
}
\value{
a list with components
\itemize{
\item{\code{lp}} {a named list containing the linear predictors (the names corespond to the fitted parameters, e.g. \code{mu}, \code{sigma}, \code{nu}, and \code{tau})}
\item{\code{fv}} {a named list containing the fitted values (the names corespond to the fitted parameters, e.g. \code{mu}, \code{sigma}, \code{nu}, and \code{tau})}
}
}
\description{
Using the object generated by \code{\link{gamlssReport}}, it makes predictions (for the parameters), based on the new data
}
\details{
For factors enter a value; this value must be an existing level of the factor. For the reference value enter \code{"reference"}.
}
\examples{
library(gamlss)


data(aids)
aids1<-gamlss(y~pb(x,df=4)+qrt,data=aids,family=PO)
m.aids1<-gamlssReport(aids1)
new_data<-data.frame(x=c(10,15,2,4),qrt=c("reference",3,4,2) )
predict.gamlssReport(m.aids1,newdata=new_data   )
}
\seealso{
\code{\link{gamlssReport}}
}
\author{
Rok Blagus, \email{rok.blagus@mf.uni-lj.si}
}
